Refactor: clean up console runner, remove preview, add exit codes for errors
